Get species information from the EASIN's Catalogue Web Service

This function retrieves species information from the EASIN's Catalogue. Users can retrieve records by species’ scientific name, environment, impact, taxonomy, Union concern status (LegalFramework). More on EASIN Web Services.

Usage

get_species(
  easin_id = NULL,
  scientific_name = NULL,
  environment = NULL,
  country_code = NULL,
  region_code = NULL,
  impact = NULL,
  taxonomy = NULL,
  union_concern = NULL
)

Arguments

easin_id

Integer. EASIN Species ID(s).

scientific_name

Character. Scientific name(s) or part(s) of it. Case insensitive.

environment

Character. Environment type(s): one or more from "MAR", "FRW", "TER", "OLI" to filter species by, marine, freshwater, terrestrial or oligohaline environments respectively.

country_code

Character. Countries' ISO 3166-1 alpha-2 code(s) to filter species of Member State concern. Use countries() to look up the list of codes. Source: EASIN Catalogue Web Service documentation. Only few states submitted their species of Member State concern to EASIN.

region_code

Character. Species of Outermost regions concern codes as defined in NUTS (Nomenclature of territorial units for statistics). Use regions() to look up the list of codes. Source: EASIN Catalogue Web Service documentation.

impact

Character. Species impact. One of "hi" (high) or "lo" (low).

taxonomy

Character named vector with the taxonomic names named by their taxonomic rank. Provide them in the right order from kingdom up to family. Source: EASIN Catalogue Web Service documentation.

union_concern

Logical. If TRUE, returns only species of Union concern. Only TRUE is allowed.

taxon

Character named vector of length 1 with the taxon name named by its taxonomic rank. Use ranks() to look up the list of valid ranks. Source: EASIN Catalogue Web Service documentation.

Value

A tibble data frame containing species information.

Details

List of supported Outermost region codes (argument region_code):

ES7: Canary Islands

FRY1: Guadeloupe

FRY2: Martinique

FRY3: French Guiana

FRY4: Réunion

FRY5: Mayotte

FRY6: Saint Martin

PT2: Azores

PT3: Madeira

Examples

# Get list of all species in the EASIN catalogue
get_species()
#> # A tibble: 14,353 × 12
#>    EasinID Name     Authorship LSID  Reference HasImpact IsEUConcern IsMSConcern
#>    <chr>   <chr>    <chr>      <chr> <chr>     <lgl>     <lgl>       <lgl>      
#>  1 R19422  Candida… "Fagen et… urn:… https://… FALSE     FALSE       FALSE      
#>  2 R11525  Candida… ""         urn:… https://… FALSE     FALSE       FALSE      
#>  3 R19423  Candida… ""         urn:… https://… FALSE     FALSE       FALSE      
#>  4 R19424  Candida… ""         urn:… https://… FALSE     FALSE       FALSE      
#>  5 R19425  Candida… ""         urn:… https://… FALSE     FALSE       FALSE      
#>  6 R19426  Candida… ""         urn:… https://… FALSE     FALSE       FALSE      
#>  7 R19427  Candida… ""         urn:… https://… FALSE     FALSE       FALSE      
#>  8 R19428  Candida… ""         urn:… https://… FALSE     FALSE       FALSE      
#>  9 R11526  Candida… ""         urn:… https://… TRUE      FALSE       FALSE      
#> 10 R19429  Candida… ""         urn:… https://… FALSE     FALSE       FALSE      
#> # ℹ 14,343 more rows
#> # ℹ 4 more variables: IsOutermostConcern <lgl>, IsPartNative <lgl>,
#> #   IsHorizonScanning <lgl>, Status <chr>

# Get list of all species of Union concern
get_species(union_concern = TRUE)
#> # A tibble: 114 × 25
#>    EASINID Name      Authorship FirstIntroductionsInEU PresentInCountries Status
#>    <chr>   <chr>     <chr>      <list>                 <list>             <chr> 
#>  1 R00046  Acacia m… De Wild.   <df [1 × 3]>           <df [8 × 1]>       A     
#>  2 R00053  Acacia s… (Labill.)… <df [1 × 3]>           <df [22 × 1]>      A     
#>  3 R00210  Acridoth… (Linnaeus… <df [1 × 3]>           <df [9 × 1]>       A     
#>  4 R00212  Acridoth… (Linnaeus… <df [1 × 3]>           <df [24 × 1]>      A     
#>  5 R00460  Ailanthu… (Mill.) S… <df [1 × 3]>           <df [52 × 1]>      A     
#>  6 R00644  Alopoche… (Linnaeus… <df [1 × 3]>           <df [38 × 1]>      A     
#>  7 R00669  Alternan… (Mart.) G… <df [1 × 3]>           <df [4 × 1]>       A     
#>  8 R00826  Ameiurus… (Rafinesq… <df [1 × 3]>           <df [30 × 1]>      A     
#>  9 R00994  Andropog… L.         <df [1 × 3]>           <df [3 × 1]>       A     
#> 10 R01506  Arthurde… (Dendy, 1… <df [1 × 3]>           <df [8 × 1]>       A     
#> # ℹ 104 more rows
#> # ℹ 19 more variables: HasImpact <lgl>, IsEUConcern <lgl>, EUConcernName <chr>,
#> #   IsOutermostConcern <lgl>, ConcernedOutermostRegions <list>,
#> #   IsMSConcern <lgl>, ConcernedMS <list>, IsPartNative <lgl>,
#> #   NativeRange <list>, IsHorizonScanning <lgl>, LastRevisionDate <lgl>,
#> #   Kingdom <chr>, Phylum <chr>, Class <chr>, Order <chr>, Family <chr>,
#> #   CommonNames <list>, Synonyms <list>, CBD_Pathways <list>

# Get info about one or more species by EASIN Species IDs
get_species(easin_id = c("R00460", "R12250"))
#> # A tibble: 2 × 25
#>   EASINID Name       Authorship FirstIntroductionsInEU PresentInCountries Status
#>   <chr>   <chr>      <chr>      <list>                 <list>             <chr> 
#> 1 R00460  Ailanthus… (Mill.) S… <df [1 × 3]>           <df [52 × 1]>      A     
#> 2 R12250  Procambar… (Girard, … <df [1 × 3]>           <df [24 × 1]>      A     
#> # ℹ 19 more variables: HasImpact <lgl>, IsEUConcern <lgl>, EUConcernName <lgl>,
#> #   IsOutermostConcern <lgl>, ConcernedOutermostRegions <list>,
#> #   IsMSConcern <lgl>, ConcernedMS <list>, IsPartNative <lgl>,
#> #   NativeRange <lgl>, IsHorizonScanning <lgl>, LastRevisionDate <lgl>,
#> #   Kingdom <chr>, Phylum <chr>, Class <chr>, Order <chr>, Family <chr>,
#> #   CommonNames <list>, Synonyms <list>, CBD_Pathways <list>

# Get info about one or more species by scientific names or parts of it
get_species(scientific_name = c("Aceria ambrosia", "Procambarus"))
#> # A tibble: 4 × 25
#>   EASINID Name       Authorship FirstIntroductionsInEU PresentInCountries Status
#>   <chr>   <chr>      <chr>      <list>                 <list>             <chr> 
#> 1 R16888  Aceria am… Wilson, 1… <df [1 × 3]>           <df [1 × 1]>       A     
#> 2 R12248  Procambar… Girard, 1… <df [1 × 3]>           <df [1 × 1]>       A     
#> 3 R12250  Procambar… (Girard, … <df [1 × 3]>           <df [24 × 1]>      A     
#> 4 R17660  Procambar… Lyko, 2017 <df [1 × 3]>           <df [16 × 1]>      A     
#> # ℹ 19 more variables: HasImpact <lgl>, IsEUConcern <lgl>, EUConcernName <chr>,
#> #   IsOutermostConcern <lgl>, ConcernedOutermostRegions <list>,
#> #   IsMSConcern <lgl>, ConcernedMS <list>, IsPartNative <lgl>,
#> #   NativeRange <lgl>, IsHorizonScanning <lgl>, LastRevisionDate <lgl>,
#> #   Kingdom <chr>, Phylum <chr>, Class <chr>, Order <chr>, Family <chr>,
#> #   CommonNames <list>, Synonyms <list>, CBD_Pathways <list>

# Get species by `environment`
get_species(environment = c("MAR","OLI"))
#> # A tibble: 1,861 × 25
#>    EASINID Name      Authorship FirstIntroductionsInEU PresentInCountries Status
#>    <chr>   <chr>     <chr>      <list>                 <list>             <chr> 
#>  1 R20136  Ablennes… "(Valenci… <df [1 × 3]>           <df [2 × 1]>       A     
#>  2 R20210  Abudefdu… "(Linnaeu… <df [1 × 3]>           <NULL>             A     
#>  3 R19331  Abudefdu… "(Steinda… <df [1 × 3]>           <df [2 × 1]>       A     
#>  4 R19652  Abudefdu… "(Lacepèd… <df [1 × 3]>           <df [4 × 1]>       A     
#>  5 R19332  Abudefdu… "(Quoy & … <df [1 × 3]>           <df [7 × 1]>       A     
#>  6 R16518  Abyla tr… "Quoy & G… <df [1 × 3]>           <df [2 × 1]>       Q     
#>  7 R17484  Acanthar… "(Forest,… <df [1 × 3]>           <df [1 × 1]>       A     
#>  8 R16519  Acanthas… "(Linnaeu… <df [1 × 3]>           <df [3 × 1]>       A     
#>  9 R18239  Acanthop… "(Forsskå… <df [1 × 3]>           <df [4 × 1]>       A     
#> 10 R00082  Acanthop… "(Delile)… <df [1 × 3]>           <df [10 × 1]>      C     
#> # ℹ 1,851 more rows
#> # ℹ 19 more variables: HasImpact <lgl>, IsEUConcern <lgl>, EUConcernName <chr>,
#> #   IsOutermostConcern <lgl>, ConcernedOutermostRegions <list>,
#> #   IsMSConcern <lgl>, ConcernedMS <list>, IsPartNative <lgl>,
#> #   NativeRange <list>, IsHorizonScanning <lgl>, LastRevisionDate <lgl>,
#> #   Kingdom <chr>, Phylum <chr>, Class <chr>, Order <chr>, Family <chr>,
#> #   CommonNames <list>, Synonyms <list>, CBD_Pathways <list>

# Get species by `country_code`
get_species(country_code = c("IE", "LT"))
#> # A tibble: 112 × 25
#>    EASINID Name      Authorship FirstIntroductionsInEU PresentInCountries Status
#>    <chr>   <chr>     <chr>      <list>                 <list>             <chr> 
#>  1 R17242  Alexandr… (Howell) … <df [1 × 3]>           <df [6 × 1]>       A     
#>  2 R00595  Allium t… L.         <df [1 × 3]>           <df [22 × 1]>      A     
#>  3 R17488  Amphibal… Darwin, 1… <df [1 × 3]>           <df [26 × 1]>      C     
#>  4 R01070  Anser an… (Linnaeus… <df [1 × 3]>           <df [55 × 1]>      C     
#>  5 R01255  Aponoget… L.f.       <df [1 × 3]>           <df [9 × 1]>       A     
#>  6 R01826  Azolla f… Lam.       <df [1 × 3]>           <df [35 × 1]>      A     
#>  7 R02253  Branta c… (Linnaeus… <df [1 × 3]>           <df [39 × 1]>      A     
#>  8 R19497  Bufo bufo (Linnaeus… <df [1 × 3]>           <df [42 × 1]>      A     
#>  9 R02476  Cabomba … A.Gray     <df [1 × 3]>           <df [15 × 1]>      A     
#> 10 R02720  Caprella… Schurin, … <df [1 × 3]>           <df [12 × 1]>      A     
#> # ℹ 102 more rows
#> # ℹ 19 more variables: HasImpact <lgl>, IsEUConcern <lgl>, EUConcernName <chr>,
#> #   IsOutermostConcern <lgl>, ConcernedOutermostRegions <list>,
#> #   IsMSConcern <lgl>, ConcernedMS <list>, IsPartNative <lgl>,
#> #   NativeRange <list>, IsHorizonScanning <lgl>, LastRevisionDate <lgl>,
#> #   Kingdom <chr>, Phylum <chr>, Class <chr>, Order <chr>, Family <chr>,
#> #   CommonNames <list>, Synonyms <list>, CBD_Pathways <list>

# Get species by `region_code`
get_species(region_code = c("ES7", "PT3"))
#> # A tibble: 362 × 25
#>    EASINID Name      Authorship FirstIntroductionsInEU PresentInCountries Status
#>    <chr>   <chr>     <chr>      <list>                 <list>             <chr> 
#>  1 R01047  Anolis a… Barbour &… <df [1 × 3]>           <df [1 × 1]>       Q     
#>  2 R01048  Anolis c… Voigt, 18… <df [1 × 3]>           <df [2 × 1]>       Q     
#>  3 R01049  Anolis e… Merrem, 1… <df [1 × 3]>           <df [1 × 1]>       Q     
#>  4 R01050  Anolis p… Gray, 1840 <df [1 × 3]>           <df [2 × 1]>       Q     
#>  5 R18148  Anolis s… Duméril &… <df [1 × 3]>           <df [3 × 1]>       Q     
#>  6 R01775  Austrocy… (Lam.) Ba… <NULL>                 <df [3 × 1]>       A     
#>  7 R03305  Chamaele… Duméril &… <df [1 × 3]>           <df [3 × 1]>       Q     
#>  8 R03485  Chlamydo… Gray, 1825 <df [1 × 3]>           <df [1 × 1]>       Q     
#>  9 R04627  Cyclura … (Bonnater… <df [1 × 3]>           <df [1 × 1]>       Q     
#> 10 R04773  Cytisus … (Hill) Ro… <df [1 × 3]>           <df [10 × 1]>      A     
#> # ℹ 352 more rows
#> # ℹ 19 more variables: HasImpact <lgl>, IsEUConcern <lgl>, EUConcernName <chr>,
#> #   IsOutermostConcern <lgl>, ConcernedOutermostRegions <list>,
#> #   IsMSConcern <lgl>, ConcernedMS <list>, IsPartNative <lgl>,
#> #   NativeRange <list>, IsHorizonScanning <lgl>, LastRevisionDate <lgl>,
#> #   Kingdom <chr>, Phylum <chr>, Class <chr>, Order <chr>, Family <chr>,
#> #   CommonNames <list>, Synonyms <list>, CBD_Pathways <list>

# Get species by `taxon`
get_species(taxon = c(family = "Vespidae"))
#> Error in get_species(taxon = c(family = "Vespidae")): Argument 'taxonomy' must be a named vector of length 5.

# Get species by full `taxonomy` levels (up to family)
get_species(
  taxonomy = list(
    kingdom = "Animalia",
    phylum = "Arthropoda",
    class = "Insecta",
    order = "Hymenoptera",
    family = "Vespidae"
  )
)
#> Error in get_species(taxonomy = list(kingdom = "Animalia", phylum = "Arthropoda",     class = "Insecta", order = "Hymenoptera", family = "Vespidae")): Argument 'taxonomy' must be a named vector of length 5.